add a seperate service.go file to manage all services required by the server
